requirements.txt
Django>=4.2,<5
psycopg2-binary
gunicorn
whitenoise
djangorestframework
drf-yasg
django-bootstrap4